# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Backend API
  description: Spring Boot Backend Application REST API
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local Development Server
  - url: https://api.example.com
    description: Production Server
security:
  - bearerAuth: []
paths:
...
  /api/users:
    get:
      tags:
        - User Management
      summary: Get all users
      description: Retrieve paginated list of all users
      parameters:
        - name: page
          in: query
          description: Page number (0-based)
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: Page size
          schema:
            type: integer
            default: 10
        - name: sortBy
          in: query
          description: Sort by field
          schema:
            type: string
            default: id
        - name: sortDir
          in: query
          description: Sort direction
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - name: search
          in: query
          description: Search term
          schema:
            type: string
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedUsers'
        '403':
          description: Access denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/users/{id}:
    get:
      tags:
        - User Management
      summary: Get user by ID
      description: Retrieve a specific user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - User Management
      summary: Update user
      description: Update an existing user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - User Management
      summary: Delete user
      description: Soft delete a user (mark as inactive)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: User deleted successfully
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: admin
        email:
          type: string
          format: email
          example: admin@example.com
        active:
          type: boolean
          example: true
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
        createdAt:
          type: string
          format: date-time
          example: 2023-12-19T10:00:00Z
        updatedAt:
          type: string
          format: date-time
          example: 2023-12-19T10:00:00Z
    UserUpdateRequest:
      type: object
      properties:
        username:
          type: string
          maxLength: 50
          example: newusername
        email:
          type: string
          format: email
          maxLength: 100
          example: newemail@example.com
        password:
          type: string
          format: password
          minLength: 6
          maxLength: 40
          example: newpassword123
    Role:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          enum: [ROLE_USER, ROLE_MODERATOR, ROLE_ADMIN]
          example: ROLE_ADMIN
        createdAt:
          type: string
          format: date-time
          example: 2023-12-19T10:00:00Z
...
    PagedUsers:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/User'
        pageable:
          $ref: '#/components/schemas/Pageable'
        totalElements:
          type: integer
          format: int64
          example: 100
        totalPages:
          type: integer
          example: 10
        last:
          type: boolean
          example: false
        first:
          type: boolean
          example: true
        numberOfElements:
          type: integer
          example: 10
        size:
          type: integer
          example: 10
        number:
          type: integer
          example: 0
...
    Pageable:
      type: object
      properties:
        sort:
          $ref: '#/components/schemas/Sort'
        pageNumber:
          type: integer
          example: 0
        pageSize:
          type: integer
          example: 10
        offset:
          type: integer
          format: int64
          example: 0
        paged:
          type: boolean
          example: true
        unpaged:
          type: boolean
          example: false
    Sort:
      type: object
      properties:
        sorted:
          type: boolean
          example: true
        unsorted:
          type: boolean
          example: false
        empty:
          type: boolean
          example: false
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: 2023-12-19T10:00:00Z
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        message:
          type: string
          example: Validation failed
        path:
          type: string
          example: /api/users
        errors:
          type: object
          additionalProperties:
            type: string
          example:
            username: Username is required
            email: Email format is invalid